+++ app/src/main/java/com/sopt/now/compose/ApiFactory.kt
@@ -2,6 +2,7 @@ package com.sopt.now.compose
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import com.sopt.now.BuildConfig
+import com.sopt.now.compose.service.AuthService
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
 import retrofit2.Retrofit
@@ -21,5 +22,4 @@ object ApiFactory {
 
 object ServicePool {
     val authService = ApiFactory.create<AuthService>()
-    val userService = ApiFactory.create<UserService>()
 }
