+++ app/src/main/java/com/sopt/now/compose/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.main
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -6,6 +6,9 @@ import androidx.activity.compose.setContent
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.sopt.now.compose.Scaffold
+import com.sopt.now.compose.signIn.SignInScreen
+import com.sopt.now.compose.signUp.SignUpScreen
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
