+++ app/src/main/java/com/sopt/now/compose/main/home/HomeScreen.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose.home
+package com.sopt.now.compose.main.home
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.lazy.LazyColumn
@@ -6,7 +6,6 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.sopt.now.compose.R
 
 @Preview
 @Composable
