+++ app/src/main/java/com/sopt/now/compose/main/home/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose.home
+package com.sopt.now.compose.main.home
 
 import com.sopt.now.compose.R
