+++ app/src/main/java/com/sopt/now/compose/request/RequestSignInDto.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
