+++ app/src/main/java/com/sopt/now/compose/response/ResponseUserInfoDto.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
